Change request unique id random generator to a better implementation
On our testing we found out that the previous random string generator
wasn't random enough. This should change as we now use the
crypto.pseudoRandomBytes call to get 12 bytes of pseudo random
string, which is converted into base64.

Notice that the base64 encoding has a few extra characters like '/'
and '+' than what the previous implementation had.